Cast update due to the use of G++ flag -Wconversion and -Wfatal-errors
git-svn-id: http://130.66.124.6/svn/maroff/Product/Cpp/Tools/random_data_generator@1705 300066c3-8af3-463d-8cc0-c0233cc64ac1

+++ random_data_generator/src/DataGenerator.cpp
@@ -15,7 +15,7 @@ template <typename T> bool outside_forbidden_zone(const T& forbidden_min, const 
 
 DataGenerator::DataGenerator(const size_t& seed) : negated(false)
 {
-    srand(seed);
+    srand((unsigned int)(seed));
 }
 
 DataGenerator& DataGenerator::no()
